moved _read_rt_lookup_file to ident base parser as it causes flashlfq conflicts for multiple input file szenarios and is not used in implemented quant parsers. fixed various test related issues.
